implement asset-based content population
this feature provides clients the ability to populate (currently, just) a node with the content extracted from an asset such as a document.

currently implemented PDF support, which will attempt to extract the text content from all pages then use that to set the content field of the target node (if any) which will also trigger a summariser/indexing job with the new content.

still to do:
- asynchronously load the pdf toolkit as it's quite slow due to WASM
- implement content prepend and append so you can upload a document without wiping the entire contents
- look into using the same architecture for the existing link fetch/hydration (echo what the frontend already does with bulk node creation)
- clean up TODOs
- ensure assets are loaded into memory using a buffer sized to the claimed Content-Length header and store this in the DB with the asset metadata