fix a bunch of UX and API issues with library page tree
now, all items are rendered in a single tree, published and draft are interspersed for easier management, drafts/submissions are kept in the correct position in the tree so authors can find them. drafts/etc are highlighted using a coloured border for easier identification of non-published items.

new implementation and tests for visibility filtering