fix OAuth2 redirect URI construction
Still not done via frontend but that's a larger piece of work. For now, we're assuming the frontend URL structure (path) and building the OAuth2 redirect URL on the backend. This means custom frontends must currently satisfy:

```
/auth/{provider}/callback
```

However in a future change this will be fully customisable by the client so custom frontends can use their own URL structure for OAuth callbacks. Or, the OAuth callback can be direct to the API. Undecided so we'll see.