reworks edit mode layout
Instead of a left-hand gutter like Notion, puts block controls above each block.

The issue with this approach is when entering edit mode, it causes a ton of layout shift as every block gets about 32-40 pixels added on top for the control bar. So I'm not super convinced this is the way forward.

The only other options are:

- A hover state menu (not mobile friendly, but Notion also doesn't provide this functionality on mobile either so...)
- A larger gutter built in to the underlying navigation layout grid, which would squish the entire site's layout but provide space on the left and right for gutter controls such as drag handles and block menus.